requirements.txt
telethon>=1.34.0
PyYAML>=6.0
httpx>=0.27.0

